requirements.txt
appdirs==1.4.0
appnope==0.1.0
cffi==1.9.1
colorful==0.1.3
coverage==4.3.4
cryptography==1.7.2
decorator==4.0.11
docopt==0.6.2
expects==0.8.0
idna==2.2
ipython==5.2.2
ipython-genutils==0.1.0
lxml==3.7.2
packaging==16.8
paramiko==2.1.1
parse==1.6.6
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.2.2
pycparser==2.17
Pygments==2.2.0
pyparsing==2.1.10
pysingleton==0.1.0
PyYAML==3.12
radish-bdd==0.2.11
redis==2.10.5
simplegeneric==0.8.1
six==1.10.0
sshtunnel==0.1.2
traitlets==4.3.1
wcwidth==0.1.7
